python/risk_scoring/requirements.txt
fastapi
uvicorn[standard]
pydantic
sentence-transformers
faiss-cpu
numpy
